\name{guess.name}
\alias{guess.name}
\title{Pluck name from a hierarchy of candidates; viz.}
\usage{guess.name(partitum)}
\description{Pluck name from a hierarchy of candidates; viz. name,
assignee, S4class, S4method, S4generic.}
\value{The guessed name (possibly \code{NULL})}
\arguments{\item{partitum}{the parsed elements}}
