\name{is.assignment}
\alias{is.assignment}
\title{Whether the expression implies assignment by <-...}
\usage{is.assignment(expression)}
\description{Whether the expression implies assignment by \code{<-}
or \code{=}.}
\value{Whether or not the expression assigns by \code{<-}
\code{=}}
\arguments{\item{expression}{the expression to check for assignment}}
