\name{parse.description}
\alias{parse.description}
\title{Parse description: the premier part of a roxygen block...}
\usage{parse.description(expression)}
\description{Parse description: the premier part of a roxygen block
containing description and option details separated by
a blank roxygen line.}
\value{A list containing the parsed description}
\arguments{\item{expression}{the description to be parsed}}
